💥 Remove deprecated traces feature
