+++ packages/datadog_flutter_plugin/android/src/main/kotlin/com/datadoghq/flutter/DatadogLogsPlugin.kt
@@ -162,7 +162,6 @@ class DatadogLogsPlugin : MethodChannel.MethodCallHandler {
             .setDatadogLogsEnabled(configuration.sendLogsToDatadog)
             .setLogcatLogsEnabled(configuration.printLogsToConsole)
             .setNetworkInfoEnabled(configuration.sendNetworkInfo)
-            .setBundleWithTraceEnabled(configuration.bundleWithTraces)
             .setBundleWithRumEnabled(configuration.bundleWithRum)
 
         configuration.loggerName?.let {
