+++ packages/datadog_flutter_plugin/integration_test_app/lib/auto_integration_scenarios/main.dart
@@ -56,9 +56,6 @@ Future<void> main() async {
       sendNetworkInfo: true,
       printLogsToConsole: true,
     ),
-    tracingConfiguration: TracingConfiguration(
-      sendNetworkInfo: true,
-    ),
     rumConfiguration: applicationId != null
         ? RumConfiguration(applicationId: applicationId)
         : null,
