+++ packages/datadog_flutter_plugin/integration_test_app/lib/integration_scenarios/integration_scenarios_screen.dart
@@ -8,7 +8,6 @@ import '../auto_integration_scenarios/rum_auto_instrumentation_scenario.dart';
 import 'logging_scenario.dart';
 import 'rum_manual_error_reporting_scenario.dart';
 import 'rum_manual_instrumentation_scenario.dart';
-import 'traces_scenario.dart';
 
 class IntegrationScenariosScreen extends StatefulWidget {
   const IntegrationScenariosScreen({Key? key}) : super(key: key);
@@ -31,7 +30,6 @@ class _IntegrationScenariosScreenState
     extends State<IntegrationScenariosScreen> {
   final items = <ScenarioItem>[
     ScenarioItem(label: 'Logging Scenario', navItem: LoggingScenario.new),
-    ScenarioItem(label: 'Traces Scenario', navItem: TracesScenario.new),
     ScenarioItem(
         label: 'Manual RUM Scenario',
         navItem: RumManualInstrumentationScenario.new),
