+++ packages/datadog_flutter_plugin/integration_test_app/lib/main.dart
@@ -46,9 +46,6 @@ Future<void> main() async {
       sendNetworkInfo: true,
       printLogsToConsole: true,
     ),
-    tracingConfiguration: TracingConfiguration(
-      sendNetworkInfo: true,
-    ),
     rumConfiguration: applicationId != null
         ? RumConfiguration(applicationId: applicationId)
         : null,
