+++ packages/datadog_tracking_http_client/example/lib/main.dart
@@ -55,9 +55,6 @@ Future<void> main() async {
       sendNetworkInfo: true,
       printLogsToConsole: true,
     ),
-    tracingConfiguration: TracingConfiguration(
-      sendNetworkInfo: true,
-    ),
     rumConfiguration: applicationId != null
         ? RumConfiguration(
             applicationId: applicationId,
